pyproject.toml
[project]
name = "fastapi-pydantic2"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Aurélien Dentan", email = "aurelien.dentan@gmail.com" }
]
dependencies = [
    "fastapi[all]>=0.100.0",
    "inject>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.19",
    "alembic>=1.11.1",
    "psycopg2>=2.9.6",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.28.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.7.0",
    "mypy>=1.4.1",
    "ruff>=0.0.278",
    "isort>=5.12.0",
    "pre-commit>=3.3.3",
    "pytest>=7.4.0",
    "httpx>=0.24.1",
]

[tool.rye.scripts]
install = { chain = ["install:precommit" ] }
"install:precommit" = "pre-commit install"
lint = { chain = ["lint:black", "lint:ruff", "lint:mypy", "lint:isort" ] }
"lint:ruff" = "ruff check src"
"lint:black" = "black --check src"
"lint:mypy" = "mypy src"
"lint:isort" = "isort src -c"
format = { chain = ["format:isort", "format:black"] }
"format:black" = "black src"
"format:isort" = "isort src"
"dev" = "uvicorn src.app.main:app --reload"
"tests" = "pytest src/tests"
"sql:autogenerate" = "alembic revision --autogenerate"
"sql:upgrade" = "alembic upgrade head"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F", "I"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"
